sat-benchmark-react/backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.12.0

# SAT Solver utilities
psutil==5.9.8
pyyaml==6.0.1

# HTTP Client (for AI integration)
httpx==0.26.0

# WebSocket support
websockets==12.0

# Background tasks
celery==5.3.6
redis==5.0.1

# CORS
python-jose[cryptography]==3.3.0

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
